feat: home page and look for the result of the llm
